Exercise You! New Update
Changed the idea from Cafe Board to an webapp which selects workouts
for you to do. This update adds the following:

1- script.js file which houses call to the exercise API I'm using from
RapidAPI

2- Skeleton HTML page for Title, Subheading, selection of exercise,
and outputing information about selected exercise below

3- Use of async/await for making call to API

4- Utilizing Javascript to create HTML elements including the heading
and Subheading

5- Utilization of javascript to popular selection drop down via
iterating through array of exercises obtained via exercise API get
request using fetch

Currently this is all this update brings. Soon it will allow users to
select an exercise from the drop down list and see what it's all about
on the page.

*Sidenote...this is really cool! I got very excited being able to
actually make this work even in its current state! CSS will be added
to make the page look better in the very near future (as in "days").